requirements.txt
# Django
Django>=5.2,<6.0
django-cors-headers>=4.3,<5.0

# OpenAI
openai>=1.0,<2.0

# LangChain (for RAG and agents)
langchain>=0.1,<1.0
langchain-openai>=0.1,<1.0
langgraph>=0.0.50,<1.0

# Vector database
pinecone>=3.0,<6.0

# Utilities
pydantic>=2.0,<3.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
